third_party/blink/web_tests/fast/forms/label/continous-click-on-label-expected.txt

Test the continuous checking and unchecking of checkbox when clicking on associated label

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkbox.checked is false
The checkbox should be checked after three clicks
PASS checkbox.checked is true
Checkbox should still be checked after ten clicks
PASS checkbox.checked is true
PASS successfullyParsed is true

TEST COMPLETE
